+++ react/Price.js
@@ -1,0 +1,3 @@
+import Price from './components/product-details/src/Price'
+
+export default Price