+++ react/ProductDetails.js
@@ -1,6 +1,0 @@
-import * as ProductDetails from './components/product-details/src'
-
-export default {
-  ...ProductDetails,
-}
-