+++ react/ProductImages.js
@@ -1,0 +1,3 @@
+import ProductImages from './components/product-details/src/ProductImages'
+
+export default ProductImages