+++ react/ShippingSimulator.js
@@ -1,0 +1,3 @@
+import ShippingSimulator from './components/product-details/src/ShippingSimulator'
+
+export default ShippingSimulator